Simplified the function so it only takes a single tract as input; added some output tracking.
